// server/lib/hxTorquescript.d.ts
export class expr_BreakStmt extends expr_Stmt {
	constructor(lineNo: number);
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_ContinueStmt extends expr_Stmt {
	constructor(lineNo: number);
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}
// ...
export class expr_ParenthesisExpr extends expr_Expr {
	constructor(expr: expr_Expr);
	expr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_ReturnStmt extends expr_Stmt {
	constructor(lineNo: number, expr: expr_Expr);
	expr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_IfStmt extends expr_Stmt {
	constructor(lineNo: number, condition: expr_Expr, body: expr_Stmt[], elseBlock: expr_Stmt[]);
	condition: expr_Expr;
	body: expr_Stmt[];
	elseBlock: expr_Stmt[];
	integer: boolean;
	endifOffset: number;
	elseOffset: number;
	ifStmtList: boolean;
	elseStmtList: boolean;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_LoopStmt extends expr_Stmt {
	constructor(lineNo: number, condition: expr_Expr, init: expr_Expr, end: expr_Expr, body: expr_Stmt[]);
	_hx_constructor(lineNo: number, condition: expr_Expr, init: expr_Expr, end: expr_Expr, body: expr_Stmt[]): void;
	condition: expr_Expr;
	init: expr_Expr;
	end: expr_Expr;
	body: expr_Stmt[];
	integer: boolean;
	loopBlockStartOffset: number;
	continueOffset: number;
	breakOffset: number;
	isForLoop: boolean;
	isStatementList: boolean;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class Token {
	constructor(type: any, lexeme: string, literal: any, line: number, position: number);
	type: any;
	lexeme: string;
	literal: any;
	line: number;
	position: number;
}

export class expr_BinaryExpr extends expr_Expr {
	constructor(lineNo: number);
	_hx_constructor(lineNo: number): void;
	left: expr_Expr;
	right: expr_Expr;
	op: Token;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
}

export class expr_FloatBinaryExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}

export class expr_IntBinaryExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	subType: any;
	operand: expr_OpCode;
	getSubTypeOperand(): void;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}
// ...
export class expr_CommaCatExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr);
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
}

export class expr_ConditionalExpr extends expr_Expr {
	constructor(condition: expr_Expr, trueExpr: expr_Expr, falseExpr: expr_Expr);
	condition: expr_Expr;
	trueExpr: expr_Expr;
	falseExpr: expr_Expr;
	integer: boolean;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_IntUnaryExpr extends expr_Expr {
	constructor(expr: expr_Expr, op: Token);
	_hx_constructor(expr: expr_Expr, op: Token): void;
	expr: expr_Expr;
	op: Token;
	integer: boolean;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_FloatUnaryExpr extends expr_Expr {
	constructor(expr: expr_Expr, op: Token);
	_hx_constructor(expr: expr_Expr, op: Token): void;
	expr: expr_Expr;
	op: Token;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}
// ...
export class expr_IntExpr extends expr_Expr {
	constructor(lineNo: number, value: number);
	value: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_FloatExpr extends expr_Expr {
	constructor(lineNo: number, value: number);
	value: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}
// ...
export class expr_ConstantExpr extends expr_Expr {
	constructor(name: Token);
	name: Token;
	fVal: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_AssignExpr extends expr_Expr {
	constructor(varExpr: expr_VarExpr, expr: expr_Expr);
	varExpr: expr_VarExpr;
	expr: expr_Expr;
	subType: any;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_AssignOpExpr extends expr_Expr {
	constructor(varExpr: expr_VarExpr, expr: expr_Expr, op: Token);
	varExpr: expr_VarExpr;
	expr: expr_Expr;
	op: Token;
	subType: any;
	operand: expr_OpCode;
	getAssignOpTypeOp(): void;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export const enum expr_FuncCallType {
	FunctionCall = 0,
	MethodCall = 1,
	ParentCall = 2,
}

export class expr_FuncCallExpr extends expr_Expr {
	constructor(name: Token, namespace: Token, args: expr_Expr[], callType: expr_FuncCallType);
	name: Token;
	namespace: Token;
	args: expr_Expr[];
	callType: expr_FuncCallType;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_SlotAccessExpr extends expr_Expr {
	constructor(objectExpr: expr_Expr, arrayExpr: expr_Expr, slotName: Token);
	objectExpr: expr_Expr;
	arrayExpr: expr_Expr;
	slotName: Token;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_SlotAssignExpr extends expr_Expr {
	constructor(objectExpr: expr_Expr, arrayExpr: expr_Expr, slotName: Token, expr: expr_Expr);
	objectExpr: expr_Expr;
	arrayExpr: expr_Expr;
	slotName: Token;
	expr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_SlotAssignOpExpr extends expr_Expr {
	constructor(objectExpr: expr_Expr, arrayExpr: expr_Expr, slotName: Token, expr: expr_Expr, op: Token);
	objectExpr: expr_Expr;
	arrayExpr: expr_Expr;
	slotName: Token;
	expr: expr_Expr;
	op: Token;
	subType: any;
	operand: expr_OpCode;
	getAssignOpTypeOp(): void;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_ObjectDeclExpr extends expr_Expr {
	constructor(className: expr_Expr, parentObject: Token, objectNameExpr: expr_Expr, args: expr_Expr[], slotDecls: expr_SlotAssignExpr[], subObjects: expr_ObjectDeclExpr[], structDecl: boolean);
	_hx_constructor(className: expr_Expr, parentObject: Token, objectNameExpr: expr_Expr, args: expr_Expr[], slotDecls: expr_SlotAssignExpr[], subObjects: expr_ObjectDeclExpr[], structDecl: boolean): void;
	className: expr_Expr;
	parentObject: Token;
	objectNameExpr: expr_Expr;
	args: expr_Expr[];
	slotDecls: expr_SlotAssignExpr[];
	subObjects: expr_ObjectDeclExpr[];
	structDecl: boolean;
	failOffset: number;
	print(indent: number, isStmt: boolean): string;
	precompileSubObject(compiler: Compiler, typeReq: any): number;
	compileSubObject(compiler: Compiler, context: CompileContext, typeReq: any, root: boolean): number;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_FunctionDeclStmt extends expr_Stmt {
	constructor(functionName: Token, args: expr_VarExpr[], stmts: expr_Stmt[], namespace: Token);
	functionName: Token;
	packageName: Token;
	args: expr_VarExpr[];
	stmts: expr_Stmt[];
	namespace: Token;
	endOffset: number;
	argc: number;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}
// ...
export class expr_Stmt {
	constructor(lineNo: number);
	_hx_constructor(lineNo: number): void;
	lineNo: number;
	print(indent: number, isStmt: boolean): string;
	printIndent(indent: number): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	addBreakCount(compiler: Compiler): void;
	visitStmt(optimizerPass: IASTVisitor): void;
	addBreakLine(ip: number, compiler: Compiler, context: CompileContext): void;
	static _hx_skip_constructor: boolean;
	static recursion: number;
	static precompileBlock(compiler: Compiler, stmts: expr_Stmt[], loopCount: number): number;
	static compileBlock(compiler: Compiler, context: CompileContext, stmts: expr_Stmt[]): number;
	static visitBlock(optimizerPass: IASTVisitor, stmts: expr_Stmt[]): void;
	static printBlock(stmt: expr_Stmt[], indent: number): string;
}

export class JSGenerator {
	constructor(stmts: expr_Stmt[]);
	generate(embedLibrary: boolean): string;
	static bootstrapEmbed(): string;
}

export class Log {
	protected constructor();
	static println(text: string): void;
	static print(text: string): void;
	static setOutputFunction(func: (arg0: string, arg1: boolean) => void): void;
}

export class Parser {
	constructor(tokens: Token[]);
	parse(): expr_Stmt[];
}

export class Scanner {
	constructor(s: string);
	scanTokens(): Token[];
	static unescape(s: string): string;
	static escape(s: string): string;
}